fix(player): move the slider throttle into a framework-free module
The frontend tests run under plain `node --test` with no npm install, so a test
may not import anything that pulls in preact. sliderThrottle.test.mjs imported
throttleTrailing from Controls.js and took the whole component with it:

    Cannot find package 'preact' imported from .../components/Controls.js

It passed locally only because node_modules happened to exist there. CI has
none, and the workflow never installs any.

The project already has the answer to this: pure logic lives in .mjs modules
beside the components rather than inside them — stereoPresentation.mjs is the
existing example. The throttle now follows that, which is where it belonged
anyway: it is a general rate limiter with nothing to do with rendering.

Verified by moving node_modules aside to reproduce the CI condition exactly;
the suite goes from 1 failure to 28/28.

Refs: https://github.com/gesellix/Bose-SoundTouch/issues/699
